Closes #2807: updated FxaReceivedTabTest to match new implementation
