+++ app/src/test/java/org/mozilla/tv/firefox/fxa/FxaRepoTest.kt
@@ -17,6 +17,7 @@ import io.reactivex.subjects.PublishSubject
 import mozilla.components.concept.sync.AuthType
 import mozilla.components.concept.sync.OAuthAccount
 import mozilla.components.concept.sync.Profile
+import mozilla.components.concept.sync.TabData
 import mozilla.components.service.fxa.manager.FxaAccountManager
 import org.junit.Before
 import org.junit.BeforeClass
@@ -24,6 +25,7 @@ import org.junit.Test
 import org.mozilla.tv.firefox.helpers.RxTestHelper
 import org.mozilla.tv.firefox.telemetry.SentryIntegration
 import org.mozilla.tv.firefox.telemetry.TelemetryIntegration
+import java.util.concurrent.TimeUnit
 
 class FxaRepoTest {
 
@@ -122,5 +124,36 @@ class FxaRepoTest {
         val profile = Profile("uid", "email", null, "displayName")
         fxaRepo.accountObserver.onProfileUpdated(profile)
         accountStateTestObs.assertValueAt(1) { it is FxaRepo.AccountState.AuthenticatedWithProfile }
+    }
+
+    @Test
+    fun `WHEN a receive tab event occurs with non-blank URLs THEN telemetry records the event`() {
+        val tabEvent = mockADMTabReceivedEventWithNullDevice(listOf(TabData("TabName", url = "https://mozilla.org")))
+        receivedTabsRaw.onNext(tabEvent)
+
+        verify(exactly = 1) { telemetryIntegration.receivedTabEvent(any()) }
+    }
+
+    @Test
+    fun `WHEN fxa state changes THEN telemetry events carrying reauthentication state are sent`() {
+        fun waitPastDebounce() {
+            testScheduler.advanceTimeBy(11, TimeUnit.SECONDS)
+        }
+
+        fxaRepo.accountObserver.onAuthenticated(mockk(relaxed = true), mockk(relaxed = true))
+        waitPastDebounce()
+        verify(exactly = 1) { telemetryIntegration.doesFxaNeedReauthenticationEvent(false) }
+
+        fxaRepo.accountObserver.onLoggedOut()
+        waitPastDebounce()
+        verify(exactly = 2) { telemetryIntegration.doesFxaNeedReauthenticationEvent(false) }
+
+        fxaRepo.accountObserver.onProfileUpdated(mockk(relaxed = true))
+        waitPastDebounce()
+        verify(exactly = 3) { telemetryIntegration.doesFxaNeedReauthenticationEvent(false) }
+
+        fxaRepo.accountObserver.onAuthenticationProblems()
+        waitPastDebounce()
+        verify(exactly = 1) { telemetryIntegration.doesFxaNeedReauthenticationEvent(true) }
     }
 }
